Moved messages to CFX.Production.Assembly.PressInsertion namespace.  Moved all structures to CFX.Structures.PressInsertion namespace.  Removed StatusResultData class, since it was not used or referenced anywhere.
