+++ CFX/Structures/PressInsertion/Condition.cs
@@ -17,34 +17,41 @@ limitations under the License.
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 
-namespace CFX.Structures
+namespace CFX.Structures.PressInsertion
 {
     /// <summary>
-    /// Describes a individual step of a Condition sequence
+    /// Describes a Condition Sequence.  Conditions consist of one or more steps that control non-pressing actions for a press recipe.
     /// </summary>
-    public class ConditionStep
+    public class Condition
     {
-        /// <summary>
-        /// The sequence position of this step in relation to the other steps in the condition.  Steps are peformed in increasing sequence integer order 1...n
-        /// </summary>
-        public int SequencePosition
+        public Condition()
         {
-            get;
-            set;
+            Steps = new ConditionStep[1];
         }
 
         /// <summary>
-        /// The name of this step
+        /// The  name of the condition.
         /// </summary>
-        public string Name
+        public string ConditionName
         {
             get;
             set;
-        }
+        }//ConditionName
 
         /// <summary>
-        /// String containing the JavaScript code that controls the step action.
+        /// An array of the steps of the Condition.
+        /// </summary>
+        public ConditionStep[] Steps
+        {
+            get;
+            set;
+        }//Steps
+
+        /// <summary>
+        /// String containing the aggregated JavaScript code of all condition steps.
         /// </summary>
         public string JavaScriptCode
         {
@@ -52,5 +59,14 @@ namespace CFX.Structures
             set;
         }//JavaScriptCode
 
-    }
+        /// <summary>
+        /// Indicates whether condition is a single step or consists of multiple sub-steps
+        /// </summary>
+        public Boolean MultiStep
+        {
+            get;
+            set;
+        }//MultiStep
+
+    }//Condition
 }
